backend/requirements.txt
# ===========================
# FastAPI & Web Framework
# ===========================
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# ===========================
# LangChain & LangGraph (AI Agents)
# ===========================
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.0
langchain-ollama==0.2.0
langgraph==0.2.45
langsmith==0.1.140

# ===========================
# Configuration
# ===========================
pydantic-settings==2.5.2
python-dotenv==1.0.1

# ===========================
# Database
# ===========================
sqlalchemy==2.0.35
aiosqlite==0.20.0

# ===========================
# Testing
# ===========================
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# ===========================
# Gesture Detection & Image Processing
# ===========================
opencv-python
mediapipe
numpy==1.26.4
Pillow

# ===========================
# Vector Search (Qdrant + SigLIP)
# ===========================
qdrant-client
transformers
torch
torchvision
accelerate
sentencepiece
protobuf

# ===========================
# Web Scraping (Optional - for data collection)
# ===========================
requests
beautifulsoup4
pandas
tqdm
